annotate all required/optional fields properly
This explicitly marks fields with optional/required in a way that
matches the original "requiredness" of the API in the MCO repo. Some of
these fields should probably be changed to required --
KubeletConfigCondition members for example should not be entirely
optional, but they always have been, and so for now we're going to leave
them set how they were. Later we can come back and fix them and see what
breaks.

This does set +kubebuilder:validation:Optional, but that's because we
historically have a field in ContinerRuntimeConfigSpec that was both
required in the CRD and tagged as omitempty in types.go and for now I'm
trying to preserve that until we can figure out if it's still necessary.